Make base offset method and invalid offset long

+++ src/java.base/share/classes/java/lang/invoke/VarHandles.java
@@ -196,7 +196,7 @@ final class VarHandles {
 
         Class<?> componentType = arrayClass.getComponentType();
 
-        int aoffset = UNSAFE.arrayBaseOffset(arrayClass);
+        int aoffset = (int) UNSAFE.arrayBaseOffset(arrayClass);
         int ascale = UNSAFE.arrayIndexScale(arrayClass);
         int ashift = 31 - Integer.numberOfLeadingZeros(ascale);
